Removed code to disable Set Default and Remove buttons when list view loses focus
I had added the code because the tab order didn't appear to function correctly when using VS options dialog,
although it worked fine in my TestUI project. It turned out that the order which the controls are added affect
tab order more that setting tab order in the designer. Doh!

Anyway, this now works correctly.